_socket: extract into module/_socket/ following the time module pattern
importing.rs had grown to 14,472 lines with 46 init_* functions inlined.
The time, math, sys, _io, _sre, _weakref and __builtin__ modules already
follow the module/<name>/{mod,moduledef,interp_<name>}.rs layout that
mirrors pypy/module/<name>/{moduledef,interp_<name>}.py. _socket was
still inlined despite being one of the largest modules.

Move the entire _socket block (FFI externs, HostentRaw/ServentRaw,
idna_converter, socket_converted_error, init_socket / init_socket_getaddrinfo
/ init_socket_type plus all socket-class helpers and the SOCKET_TYPE_OBJ
/ DEFAULT_SOCKET_TIMEOUT thread-locals) into
module/_socket/interp_socket.rs. The previous `fn init_socket` is now
`pub fn register_module` and the existing inline closures are kept
verbatim. moduledef.rs is the PyPy-equivalent thin glue that calls
register_module. importing.rs now references
crate::module::_socket::moduledef::init for module registration.

3343 lines move out of importing.rs (14,472 -> 11,132).
