+++ pyre/pyre-interpreter/src/module/mod.rs
@@ -5,6 +5,8 @@
 
 pub mod __builtin__;
 pub mod _io;
+#[allow(non_snake_case)]
+pub mod _socket;
 pub mod _sre;
 pub mod _weakref;
 pub mod math;
